cloud-generator/requirements.txt
click ==8.0.3
Jinja2 ==3.0.3
google-cloud-storage ==2.2.1
google-cloud-pubsub ==2.12.0

